src/main/webapp/app/entities/photodb/file/update/file-update.component.ts
import { Component, OnInit } from '@angular/core';
import { HttpResponse } from '@angular/common/http';
import { ActivatedRoute } from '@angular/router';
import { Observable } from 'rxjs';
import { finalize, map } from 'rxjs/operators';

import { FileFormService, FileFormGroup } from './file-form.service';
import { IFile } from '../file.model';
import { FileService } from '../service/file.service';
import { IFolder } from 'app/entities/photodb/folder/folder.model';
import { FolderService } from 'app/entities/photodb/folder/service/folder.service';

@Component({
  selector: 'jhi-file-update',
  templateUrl: './file-update.component.html',
})
export class FileUpdateComponent implements OnInit {
  isSaving = false;
  file: IFile | null = null;

  foldersSharedCollection: IFolder[] = [];

  editForm: FileFormGroup = this.fileFormService.createFileFormGroup();

  constructor(
    protected fileService: FileService,
    protected fileFormService: FileFormService,
    protected folderService: FolderService,
    protected activatedRoute: ActivatedRoute
  ) {}

  compareFolder = (o1: IFolder | null, o2: IFolder | null): boolean => this.folderService.compareFolder(o1, o2);

  ngOnInit(): void {
    this.activatedRoute.data.subscribe(({ file }) => {
      this.file = file;
      if (file) {
        this.updateForm(file);
      }

      this.loadRelationshipsOptions();
    });
  }

  previousState(): void {
    window.history.back();
  }

  save(): void {
    this.isSaving = true;
    const file = this.fileFormService.getFile(this.editForm);
    if (file.id !== null) {
      this.subscribeToSaveResponse(this.fileService.update(file));
    } else {
      this.subscribeToSaveResponse(this.fileService.create(file));
    }
  }

  protected subscribeToSaveResponse(result: Observable<HttpResponse<IFile>>): void {
    result.pipe(finalize(() => this.onSaveFinalize())).subscribe({
      next: () => this.onSaveSuccess(),
      error: () => this.onSaveError(),
    });
  }

  protected onSaveSuccess(): void {
    this.previousState();
  }

  protected onSaveError(): void {
    // Api for inheritance.
  }

  protected onSaveFinalize(): void {
    this.isSaving = false;
  }

  protected updateForm(file: IFile): void {
    this.file = file;
    this.fileFormService.resetForm(this.editForm, file);

    this.foldersSharedCollection = this.folderService.addFolderToCollectionIfMissing<IFolder>(this.foldersSharedCollection, file.folder);
  }

  protected loadRelationshipsOptions(): void {
    this.folderService
      .query()
      .pipe(map((res: HttpResponse<IFolder[]>) => res.body ?? []))
      .pipe(map((folders: IFolder[]) => this.folderService.addFolderToCollectionIfMissing<IFolder>(folders, this.file?.folder)))
      .subscribe((folders: IFolder[]) => (this.foldersSharedCollection = folders));
  }
}
